Refactor RowExpression rewrite rule sets
This commit:
1) refactored translateExpressions to inherit from RowExpressionRewriteRuleSet
2) added SimplifyRowExpressionOptimizer
