+++ presto-main/src/test/java/com/facebook/presto/sql/planner/iterative/rule/TestSimplifyExpressions.java
@@ -14,20 +14,30 @@
 package com.facebook.presto.sql.planner.iterative.rule;
 
 import com.facebook.presto.metadata.MetadataManager;
+import com.facebook.presto.spi.relation.RowExpression;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
+import com.facebook.presto.spi.type.Type;
+import com.facebook.presto.sql.TestingRowExpressionTranslator;
 import com.facebook.presto.sql.parser.SqlParser;
 import com.facebook.presto.sql.planner.LiteralEncoder;
 import com.facebook.presto.sql.planner.PlanVariableAllocator;
+import com.facebook.presto.sql.planner.TypeProvider;
 import com.facebook.presto.sql.planner.VariablesExtractor;
 import com.facebook.presto.sql.tree.Expression;
 import com.facebook.presto.sql.tree.ExpressionRewriter;
 import com.facebook.presto.sql.tree.ExpressionTreeRewriter;
 import com.facebook.presto.sql.tree.LogicalBinaryExpression;
+import com.google.common.collect.Streams;
 import org.testng.annotations.Test;
 
 import java.util.Comparator;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 import static com.facebook.presto.SessionTestUtils.TEST_SESSION;
 import static com.facebook.presto.metadata.MetadataManager.createTestMetadataManager;
@@ -37,7 +47,9 @@ import static com.facebook.presto.sql.ExpressionUtils.extractPredicates;
 import static com.facebook.presto.sql.ExpressionUtils.rewriteIdentifiersToSymbolReferences;
 import static com.facebook.presto.sql.planner.iterative.rule.SimplifyExpressions.rewrite;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
+import static java.lang.String.format;
 import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toMap;
 import static org.testng.Assert.assertEquals;
 
 public class TestSimplifyExpressions
@@ -45,6 +57,10 @@ public class TestSimplifyExpressions
     private static final SqlParser SQL_PARSER = new SqlParser();
     private static final MetadataManager METADATA = createTestMetadataManager();
     private static final LiteralEncoder LITERAL_ENCODER = new LiteralEncoder(METADATA.getBlockEncodingSerde());
+    private static final Map<String, Type> TYPES = Streams.concat(
+            Stream.of("A", "B", "C", "D", "E", "F", "I", "V", "X", "Y", "Z"),
+            IntStream.range(1, 61).boxed().map(i -> format("A%s", i)))
+            .collect(toMap(Function.identity(), string -> BOOLEAN));
 
     @Test
     public void testPushesDownNegations()
@@ -67,6 +83,9 @@ public class TestSimplifyExpressions
     @Test
     public void testExtractCommonPredicates()
     {
+        assertSimplifies("TRUE", "TRUE");
+        assertSimplifies("IF(X, X, Y)", "IF(X, X, Y)");
+
         assertSimplifies("X AND Y", "X AND Y");
         assertSimplifies("X OR Y", "X OR Y");
         assertSimplifies("X AND X", "X");
@@ -86,7 +105,7 @@ public class TestSimplifyExpressions
         assertSimplifies("((X OR V) AND X) OR ((X OR V) AND V)", "X OR V");
 
         assertSimplifies("((X OR V) AND Z) OR ((X OR V) AND V)", "(X OR V) AND (Z OR V)");
-        assertSimplifies("X AND ((Y AND Z) OR (Y AND V) OR (Y AND X))", "X AND Y AND (Z OR V OR X)");
+        assertSimplifies("X AND ((Y AND Z) OR (Y AND V) OR (Y AND X))", "X AND Y AND (Z OR V OR X)", "X AND Y");
         assertSimplifies("(A AND B AND C AND D) OR (A AND B AND E) OR (A AND F)", "A AND ((B AND C AND D) OR (B AND E) OR F)");
 
         assertSimplifies("((A AND B) OR (A AND C)) AND D", "A AND (B OR C) AND D");
@@ -115,12 +134,23 @@ public class TestSimplifyExpressions
 
     private static void assertSimplifies(String expression, String expected)
     {
+        assertSimplifies(expression, expected, null);
+    }
+
+    private static void assertSimplifies(String expression, String expected, String rowExpressionExpected)
+    {
         Expression actualExpression = rewriteIdentifiersToSymbolReferences(SQL_PARSER.createExpression(expression));
         Expression expectedExpression = rewriteIdentifiersToSymbolReferences(SQL_PARSER.createExpression(expected));
         Expression rewritten = rewrite(actualExpression, TEST_SESSION, new PlanVariableAllocator(booleanVariablesFor(actualExpression)), METADATA, LITERAL_ENCODER, SQL_PARSER);
         assertEquals(
                 normalize(rewritten),
                 normalize(expectedExpression));
+        TestingRowExpressionTranslator translator = new TestingRowExpressionTranslator(METADATA);
+        RowExpression actualRowExpression = translator.translate(actualExpression, TypeProvider.viewOf(TYPES));
+        RowExpression simplifiedRowExpression = SimplifyRowExpressions.rewrite(actualRowExpression, METADATA, TEST_SESSION.toConnectorSession());
+        Expression expectedByRowExpression = Optional.ofNullable(rowExpressionExpected).map(expr -> rewriteIdentifiersToSymbolReferences(SQL_PARSER.createExpression(expr))).orElse(rewritten);
+        RowExpression simplifiedByExpression = translator.translate(expectedByRowExpression, TypeProvider.viewOf(TYPES));
+        assertEquals(simplifiedRowExpression, simplifiedByExpression);
     }
 
     private static Set<VariableReferenceExpression> booleanVariablesFor(Expression expression)
